+++ src/ABC.cpp
@@ -380,7 +380,7 @@ arma::cube generate_acceptable_data_cpp(
 
 
 // Perform ABC sampling using the stick breaking procedure, returning the acceptance ratio.
-double compute_ABC_cpp(
+Rcpp::List compute_ABC_cpp(
       const unsigned int &n_sample,
       const unsigned int &m_sample,
       const double &alpha_0,
@@ -390,7 +390,8 @@ double compute_ABC_cpp(
       const Rcpp::NumericVector &summarize_eps,
       const unsigned int max_iter,
       const double &p_norm,
-      const bool use_optimized_summary
+      const bool use_optimized_summary,
+      const bool return_distances
 ) {
 
    const unsigned int n_obs = mtx_obs.nrow();
@@ -404,7 +405,13 @@ double compute_ABC_cpp(
    }
 
    if (p != p_obs) {
-      Rcpp::stop("Error: different number of columns (supplied nu_0: %i, needed: %i)", p, p_obs);
+      Rcpp::stop("different number of columns (supplied nu_0: %i, needed: %i)", p, p_obs);
+   }
+
+   const unsigned int n_summary = get_number_summary_statistics(use_optimized_summary);
+
+   if (summarize_eps.size() != n_summary) {
+      Rcpp::stop("summarize_eps must match the number of summary statistics (supplied: %i, needed: %i)", summarize_eps.size(), n_summary);
    }
 
    // Allocate results
@@ -412,15 +419,12 @@ double compute_ABC_cpp(
    unsigned int n_accepted = 0;
 
    // Allocate distances between summary statistics
-   const unsigned int n_summary = get_number_summary_statistics(use_optimized_summary);
-   Rcpp::NumericVector vec_distances(n_summary);
+   Rcpp::NumericMatrix mtx_distances(max_iter, n_summary);
 
-   if (summarize_eps.size() != n_summary) {
-      Rcpp::stop("Error: summarize_eps must match the number of summary statistics (supplied: %i, needed: %i)", summarize_eps.size(), n_summary);
-   }
+   // Generated data
+   Rcpp::NumericMatrix mtx_gen(n*m, p);
 
-
-   // Quantile matrix
+   // Summary statistics
    Rcpp::NumericMatrix summary_obs(n_summary, p);
    Rcpp::NumericMatrix summary_gen(n_summary, p);
 
@@ -430,28 +434,36 @@ double compute_ABC_cpp(
 
       if (t % 1000 == 0) Rcpp::checkUserInterrupt();
 
-      Rcpp::NumericMatrix mtx_gen(n*m, p);
-
       mtx_gen = rdirdirgamma_beta_cpp(n, m, alpha_0, beta_0, nu_0);
 
       summary_gen = get_summary_statistics_cpp(mtx_gen(Rcpp::Range(0, n_obs - 1), _), use_optimized_summary);
 
       // Allocate distances between summary statistics
       // vec_distances = compute_distances_gen_obs_cpp(mtx_gen, mtx_obs, p_norm, use_optimized_summary);
+      mtx_distances(t,_) = compute_distances_gen_obs_cpp(mtx_gen, mtx_obs, p_norm, use_optimized_summary);
 
       for (unsigned int i = 0; i < n_summary; i++) {
 
          // Compute distances between summary statistics
-         vec_distances[i] = norm_minkowski(summary_obs(i,_) - summary_gen(i,_), p_norm);
+         mtx_distances(t,i) = norm_minkowski(summary_obs(i,_) - summary_gen(i,_), p_norm);
       }
 
 
-      if (is_true(all(vec_distances < summarize_eps))) {
+      if (is_true(all(mtx_distances(t,_) < summarize_eps))) {
          n_accepted++;
       }
    }
 
-   return ((double) n_accepted / max_iter);
+   // Return
+
+   Rcpp::List l;
+   l["n_accepted"] = n_accepted;
+   l["accept_ratio"] = (double) n_accepted / max_iter;
+
+   if (return_distances) {
+      l["d_ABC"] = mtx_distances;
+   }
+   return (l);
 }