+++ src/RcppExports.cpp
@@ -48,8 +48,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // compute_ABC_cpp
-double compute_ABC_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const Rcpp::NumericVector& summarize_eps, const unsigned int max_iter, const double& p_norm, const bool use_optimized_summary);
-RcppExport SEXP _rdirdirgamma_compute_ABC_cpp(SEXP n_sampleSEXP, SEXP m_sampleSEXP, SEXP alpha_0SEXP, SEXP beta_0SEXP, SEXP nu_0SEXP, SEXP mtx_obsSEXP, SEXP summarize_epsSEXP, SEXP max_iterSEXP, SEXP p_normSEXP, SEXP use_optimized_summarySEXP) {
+Rcpp::List compute_ABC_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const Rcpp::NumericVector& summarize_eps, const unsigned int max_iter, const double& p_norm, const bool use_optimized_summary, const bool return_distances);
+RcppExport SEXP _rdirdirgamma_compute_ABC_cpp(SEXP n_sampleSEXP, SEXP m_sampleSEXP, SEXP alpha_0SEXP, SEXP beta_0SEXP, SEXP nu_0SEXP, SEXP mtx_obsSEXP, SEXP summarize_epsSEXP, SEXP max_iterSEXP, SEXP p_normSEXP, SEXP use_optimized_summarySEXP, SEXP return_distancesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -63,7 +63,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const unsigned int >::type max_iter(max_iterSEXP);
     Rcpp::traits::input_parameter< const double& >::type p_norm(p_normSEXP);
     Rcpp::traits::input_parameter< const bool >::type use_optimized_summary(use_optimized_summarySEXP);
-    rcpp_result_gen = Rcpp::wrap(compute_ABC_cpp(n_sample, m_sample, alpha_0, beta_0, nu_0, mtx_obs, summarize_eps, max_iter, p_norm, use_optimized_summary));
+    Rcpp::traits::input_parameter< const bool >::type return_distances(return_distancesSEXP);
+    rcpp_result_gen = Rcpp::wrap(compute_ABC_cpp(n_sample, m_sample, alpha_0, beta_0, nu_0, mtx_obs, summarize_eps, max_iter, p_norm, use_optimized_summary, return_distances));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -221,7 +222,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_rdirdirgamma_sample_ABC_rdirdirgamma_beta_cpp", (DL_FUNC) &_rdirdirgamma_sample_ABC_rdirdirgamma_beta_cpp, 9},
     {"_rdirdirgamma_generate_acceptable_data_cpp", (DL_FUNC) &_rdirdirgamma_generate_acceptable_data_cpp, 11},
-    {"_rdirdirgamma_compute_ABC_cpp", (DL_FUNC) &_rdirdirgamma_compute_ABC_cpp, 10},
+    {"_rdirdirgamma_compute_ABC_cpp", (DL_FUNC) &_rdirdirgamma_compute_ABC_cpp, 11},
     {"_rdirdirgamma_compute_distances_gen_obs_cpp", (DL_FUNC) &_rdirdirgamma_compute_distances_gen_obs_cpp, 4},
     {"_rdirdirgamma_get_number_summary_statistics", (DL_FUNC) &_rdirdirgamma_get_number_summary_statistics, 1},
     {"_rdirdirgamma_get_optimized_summary_statistics_cpp", (DL_FUNC) &_rdirdirgamma_get_optimized_summary_statistics_cpp, 1},